pyproject.toml
[tool.flake8]
max-line-length = 100
max-complexity = 18
select = ['C','E','F','W','B']
ignore = ['E501','W503','E203','D100','E704']

[tool.black]
line-length = 100
target_version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.venv
  | \.vscode
)
'''

[tool.isort]
line_length = 100
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true
known_first_party = "svc,tests"
default_section = "THIRDPARTY"
lines_after_imports = -1

[tool.autoflake]
recursive = true
in-place = true
remove-unused-variables = true
remove-all-unused-imports = true

[tool.mypy]
plugins = [
    "mypy_django_plugin.main",
    "mypy_drf_plugin.main"
]

follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
ignore_missing_imports = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.django-stubs]
django_settings_module = "days_project.settings"

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:distutils",
    "ignore::DeprecationWarning:aiojobs",
    "ignore::DeprecationWarning:asyncio",
    "ignore::DeprecationWarning:sqlalchemy_utils",
    "ignore::DeprecationWarning:pydantic",
    "ignore::pydantic.PydanticDeprecatedSince20",
    "ignore:.*pkg_resources\\.declare_namespace.*:DeprecationWarning",
    "ignore::pytest.PytestCacheWarning",
]
asyncio_mode = "auto"
log_cli = true
log_level = "INFO"

[tool.poetry]
name = "days"
version = "0.1.0"
description = ""
authors = ["Tim <yaxyzza@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
environs = {extras = ["django"], version = "^11.0.0"}
pyproject-autoflake = "^1.0.2"
drf-yasg = {extras = ["validation"], version = "^1.21.7"}
pycountry = "^24.6.1"


[tool.poetry.group.dev.dependencies]
autoflake = "^2.3.1"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.11.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.4"}
djangorestframework-stubs = {extras = ["compatible-mypy"], version = "^3.15.1"}
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
